Created routine to extract dimension from fat flat table
The data creation routine doesn't work yet, I just used a manually
created spreadsheet to demonstrate.
The routine attempts to create a dimension table and replace that with
a surrogate key in the fat/fact table. Attribute list provided is
checked and if they are one to one with the key they are kept.
Some stuff to check: how it deals with blanks, how to do it for a
bunch of dimensions, ie create a wrapper that iterates and returns the
whole thing, how to force attributes (ie rather than reject
duplicates, pick the first/last and keep a count of that for reference)