Updated routine to deal with blanks and multiple values
Added an overall check on the final dimension table to ensure the keys
are still unique.
Added an option to ignore blanks and pick the attribute value when
that is the case (only one value and blank)
Added an option to ignore multiple attributes and pick the first one.
Added sort_values() to the dim table creation, possibly a nice to have
if we are talking millions of records.
Added more print commands to explain what is going on.