Fixed fuzzy matching recipe and added a new fully native implementation
Fuzzy matching recipe had a bug, it didn't consider the exact matches
when doing the one file against itself check.  I fixed that. Also I
added a new home brew recipe that doesn't require the levensthein
library, no idea how slow it is, and needs major work on the
pre-processing to get it to the level of fuzzy wuzzy but with some
care it can be done (or we pick code from the fuzzy wuzzy python
library, if not too complex. For now it is a version from rosetta code.