Cleanup and further coding
Added the token sort ratio preprocessing so it is now a bit more like
the fuzzy wuzzy library. I removed one of the implementations for
simplicity.
TODO: benchmark how these various implementations work and pick the
fastest.
TODO: create a full recipe that also does the flattening of matches
and linking to the original key etc.